+++ config/splat.us.stst0.yaml
@@ -51,10 +51,10 @@ segments:
       - [0x27B00, .rodata, 2B0C8] # EntityStageTitleCard
       - [0x27B18, .rodata, 2B0C8] # func_801ABBBC
       - [0x27B40, .rodata, 2C564] # EntityDracula
-      - [0x27B7C, .rodata, 2C564] # "charal %x\n"
-      - [0x27B8C, .rodata, 2C564] # EntityDraculaFinalForm
-      - [0x27BDC, .rodata, 2C564] # "can't get effect works!\n"
-      - [0x27BFC, .rodata, 2C564] # func_801AF774
+      - [0x27B7C, .rodata, 2DAC8] # "charal %x\n"
+      - [0x27B8C, .rodata, 2DAC8] # EntityDraculaFinalForm
+      - [0x27BDC, .rodata, 2DAC8] # "can't get effect works!\n"
+      - [0x27BFC, .rodata, 2DAC8] # func_801AF774
       - [0x27C1C, rodata]
       - [0x27C8C, .rodata, 36358]
       - [0x27CAC, rodata]
@@ -64,6 +64,7 @@ segments:
       - [0x2A8DC, c]
       - [0x2B0C8, c]
       - [0x2C564, c] # EntityDracula
+      - [0x2DAC8, c]
       - [0x30030, c]
       - [0x3186C, c, random]
       - [0x3189C, c, update]
