+++ config/symbols.us.stst0.txt
@@ -41,13 +41,11 @@ EntityRedDoor = 0x801B3CD0;
 DestroyEntity = 0x801B4908;
 DestroyEntitiesFromIndex = 0x801B4974;
 PreventEntityFromRespawning = 0x801B49F0;
-AnimateEntity = 0x801B4A38;
 GetSideToPlayer = 0x801B4C78;
 MoveEntity = 0x801B4CBC;
 FallEntity = 0x801B4CEC;
 AllocEntity = 0x801B53CC;
 SetStep = 0x801B5794;
-SetSubStep = 0x801B57B4;
 InitializeEntity = 0x801B584C;
 EntityDummy = 0x801B5948;
 EntityPrizeDrop = 0x801B6358;
